Revert "feat: Arbitrary index function to support data splits"
This reverts commit 415ed87dae90dddcd354860bcadd4bff3364f698.
